WIP: added icon and separated classes
feature/150167-datepicker
